+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   - `outfit-wars registrations`
   - `outfit-wars matches`
   - `rolemenu list-menus`
+  - `apex craftables`
+  - `apex maps`
+  - `apex player`
 - Added the `News` feed.
 - Spiffed up the role menu commands, and added support for emojis on role labels.
 - Base capture notifications now show the amount of outfit resource earned.